actualizacion de cliente

+++ API/Entities/Client.cs
@@ -11,7 +11,7 @@ namespace API.Entities
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
-        public double Cellphone { get; set; }
+        public int Cellphone { get; set; }
         public List<Booking> TravelPackages { get; set; }
     }
 }
